Fail early + Use dask
