+++ jupyter_tiler/titiler/_server.py
@@ -10,6 +10,7 @@ from titiler.core.algorithm.base import BaseAlgorithm
 from titiler.core.dependencies import DefaultDependency
 from titiler.core.errors import DEFAULT_STATUS_CODES, add_exception_handlers
 from titiler.core.factory import TilerFactory
+from rasterio.enums import Resampling
 from xarray import DataArray
 
 from jupyter_tiler._base_server import _FastApiTileServer
@@ -18,6 +19,20 @@ from jupyter_tiler.constants._messages import (
     _not_initialized_message,
 )
 from jupyter_tiler.titiler._xarray_stac_backend import XarraySTACTilerFactory
+
+
+def _normalize_resampling(value: str | Resampling) -> Resampling:
+    """Normalize a resampling value to a rasterio Resampling enum."""
+    if isinstance(value, Resampling):
+        return value
+
+    try:
+        return Resampling[value]
+    except KeyError as e:
+        valid = ", ".join(Resampling.__members__.keys())
+        raise ValueError(
+            f"Invalid resampling '{value}'. Valid values: {valid}"
+        ) from e
 
 
 class TiTilerServer(_FastApiTileServer):
@@ -99,6 +114,7 @@ class TiTilerServer(_FastApiTileServer):
             raise ValueError("resolution_scale must be > 0")
         if viewport_width < 0 or viewport_height < 0:
             raise ValueError("viewport_width and viewport_height must be >= 0")
+        resampling = _normalize_resampling(resampling)
 
         await self.start()
